Add permissions checks for authenticated user

+++ VueApp1.Server/Controllers/UserController.cs
@@ -18,6 +18,12 @@ public class UserController(IUserRepository repository, IMapper mapper) : Contro
     [HttpGet]
     public IActionResult GetUsers([FromQuery] UserParameters query)
     {
+        var authUser = (User)HttpContext.Items["User"];
+        // Only Admins are allowed to access all users' infos
+        if (authUser.AccountType != Models.Enums.Account.Admin)
+        {
+            return Unauthorized();
+        }
         if (query.Page <= 0 || query.Limit <= 0)
         {
             return BadRequest("Error: invalid page and limit parameters!");
@@ -29,6 +35,12 @@ public class UserController(IUserRepository repository, IMapper mapper) : Contro
     [HttpGet("{id}")]
     public async Task<IActionResult> GetUser(int id)
     {
+        var authUser = (User)HttpContext.Items["User"];
+        // Only Admins are allowed to access others' infos. Else user only allowed to access own info
+        if (authUser.ID != id && authUser.AccountType != Models.Enums.Account.Admin)
+        {
+            return Unauthorized();
+        }
         var user = await repository.GetByIdAsync(id);
         if (user == null)
         {
@@ -40,13 +52,25 @@ public class UserController(IUserRepository repository, IMapper mapper) : Contro
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateUser(int id, [FromBody] UpdateRequest user)
     {
-        await repository.UpdateAsync(id, user);
-        return Ok("User successfully updated!");
+        var authUser = (User)HttpContext.Items["User"];
+        // Only Admins are allowed to change other people's user info. Else user only allowed to change own info
+        if (authUser.ID != id && authUser.AccountType != Models.Enums.Account.Admin)
+        {
+            return Unauthorized();
+        }
+        var updatedUser = await repository.UpdateAsync(id, user);
+        return Ok(updatedUser);
     }
 
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteUser(int id)
     {
+        var authUser = (User)HttpContext.Items["User"];
+        // Only Admins are allowed to delete users
+        if (authUser.AccountType != Models.Enums.Account.Admin)
+        {
+            return Unauthorized();
+        }
         await repository.DeleteAsync(id);
         return Ok("User successfully deleted!");
     }
